Move IEnumerable.Flatten to the right place
